[feat] 모임통장 거래내역 조회 API 구현

+++ src/main/java/org/hana/wooahhanaapi/domain/community/dto/CommunityTrsfRecordReqDto.java
@@ -17,5 +17,5 @@ import java.util.UUID;
 public class CommunityTrsfRecordReqDto {
 
     private UUID communityId;
-    private Long recentMonth;
+    private int recentMonth;
 }
